Refactor the ROS2 driver's module update
We now explicitly use each gripper's `start_module_updates()` instead of
calling that implicitly in the `connect()` method.

This is cleaner with the new approach where `connect()` is allowed to
fail for grippers that are not available but will connect later.

Also use a slightly higher timeout for Modbus.
This seems to be more robust with the BKS simulator.