Clean-up the test setup for parameters
It now runs successfully in conjunction with the `test_configuration`
module which caused problems before.
It seems that _global_ python code does not get executed right before
the tests but rather earlier somehow and can clash with previous tests
from other modules.
Putting the setup in the first test solves this.